+++ language_tags/tags.py
@@ -10,7 +10,7 @@ from language_tags.Tag import Tag
 
 
 parent_dir = os.path.dirname(__file__)
-data_dir = '../node_modules/language-subtag-registry/data/json/'
+data_dir = 'data/json/'
 index = json.load(open(os.path.join(parent_dir, data_dir + "index.json"), encoding='utf-8'))
 registry = json.load(open(os.path.join(parent_dir, data_dir + "registry.json"), encoding='utf-8'))
 
@@ -145,7 +145,7 @@ class tags():
         results = []
 
         macrolanguage = macrolanguage.lower()
-        macrolanguage_data = json.load(open(os.path.join(parent_dir, data_dir + "macrolanguage.json")))
+        macrolanguage_data = json.load(open(os.path.join(parent_dir, data_dir, "macrolanguage.json")))
         if macrolanguage not in macrolanguage_data:
             raise Exception('\'' + macrolanguage + '\' is not a macrolanguage.')
         for registry_item in registry:
@@ -199,5 +199,5 @@ class tags():
 
         :return: date as string (for example: '2014-03-27').
         """
-        meta = json.load(open(os.path.join(parent_dir, data_dir + "meta.json")))
+        meta = json.load(open(os.path.join(parent_dir, data_dir, "meta.json")))
         return meta['File-Date']